Update test suite and test directly against ActiveRecord
This commit reconfigures the test suite to test directly against
ActiveRecord. It accomplishes this by setting up an in-memory SQLite
store. There are several models added for each test module.

This commit also updates the RSpec configuration to be reflect the
modern DSL. This includes changes to the .rspec file and
`spec_helper.rb`.

Finally, this commit removes a lot of tests that were directly or
indirectly testing either ActiveRecord functionality (through the use of
doubles/mocks) or other module functionality (which is redundant, since
those modules are tested directly).

- Add SQLite to development dependencies
- Modify `.rspec` file to include `spec_helper.rb` in every spec
- Add a lot of models for the various specs to use and a schema, which
  ActiveRecord loads
- Update specs to use the new AR-backed models instead of mocks
- Update specs to remove redundant and/or unnecessary tests
- Update specs to new RSpec syntax/DSL