Use `require` instead of `require_dependency`
refs #45, where ActiveStorage has a circular dependency which fails to
load in one ordering (but not in the other).

As far as I can tell, `require_dependency` is mostly useful for
reloading, and since the entire use case in
`lib/consistency_fail/models.rb` is about *preloading*, this doesn't
seem necessary. Maybe this was necessary back when I was using
`ObjectSpace` as a terrible hack? Anyway, `require` it is.

This also sorts dependencies in any given `LOAD_PATH` directory before
loading them in sequence, just so whatever issues happen along these
lines in the future are more easily repeatable.